many: isolate container functionality in its own daemon (#32)
* daemons: add initial code for CVD

* many: more changes to cvd

* daemons/cvd: a bit more structuring

* wip: something

* many: add initial complete version of cvd, removed setgid/setuid from bake as we are moving to cvd

* libs/kitchen: working on migrating from cv to cvd

* many: work in progress to move a lot of the build steps into bakectl - as well as splitting up kitchen library

* many: move packing into bake

* many: more work on packing

* many: more move towards no shared kitchen

* many: more work on restructuring and moving stuff between bake and bakectl

* many: some cleanup of the refactoring - we still need to solve the path stuff for bakectl

* many: refactor how we structure the build and pack helpers

* many: we're finally back to compilation success!

* many: fix location of spaces, and correctly send the rootfs of the ns

* Casts return value to `char*` to resolve type mismatch

Adds an explicit cast to `char*` for the return value of `chef_dirs_rootfs` to address a type mismatch, ensuring compatibility with the `rootfs` variable and preventing potential compiler warnings or errors.

* libs: make sure we write default config for cvd

* cvd: couple of crash fixes

* many: switch to abstract sockets, there is still an issue with gracht where the bind_address_length is not long enough for clients as we set it to the length of the bind address

* libs: update gracht to fix abstract sockets

* many: fix issues with abstract sockets, fix a crash

* libs: fix another crash, improve output

* libs/cvd: fix snap build

* many: work on the fridge backend to allow not using chef-client

* many: add remaining fridge changes

* tools/bake/commands: fix use of VLOG_* here when not included

* daemons/cvd/server: support rootfs image download instead of debootstrap

* many: fix bakectl working directory

* tools,libs: fix snap build

* many: multiple fixes to get building in the container working

* libs/cvd: fix wrong install path

* libs/package/resolvers: fix resolving paths from ld.so

+++ daemons/waiterd/server/init.c
@@ -28,7 +28,12 @@
 #include <arpa/inet.h>
 #include <sys/un.h>
 
-static int __local_size(void) {
+static int __local_size(const char* address) {
+    // If the address starts with '@', it is an abstract socket path.
+    // Abstract socket paths are not null-terminated, so we need to account for that.
+    if (address[0] == '@') {
+        return offsetof(struct sockaddr_un, sun_path) + strlen(address);
+    }
     return sizeof(struct sockaddr_un);
 }
 
@@ -36,13 +41,25 @@ static int __configure_local(struct sockaddr_storage* storage, const char* addre
 {
     struct sockaddr_un* local = (struct sockaddr_un*)storage;
 
-    // ensure it doesn't exist
-    if (unlink(address) && errno != ENOENT) {
+    local->sun_family = AF_LOCAL;
+
+    // sanitize the address length
+    if (strlen(address) >= sizeof(local->sun_path)) {
+        fprintf(stderr, "__configure_local: address too long for local socket: %s\n", address);
         return -1;
     }
 
-    local->sun_family = AF_LOCAL;
-    strncpy(local->sun_path, address, sizeof(local->sun_path));
+    // handle abstract socket paths
+    if (address[0] == '@') {
+        local->sun_path[0] = '\0';
+        strncpy(local->sun_path + 1, address + 1, sizeof(local->sun_path) - 1);
+    } else {
+        // ensure it doesn't exist
+        if (unlink(address) && errno != ENOENT) {
+            return -1;
+        }
+        strncpy(local->sun_path, address, sizeof(local->sun_path));
+    }
     return 0;
 }
 #elif defined(_WIN32)
@@ -51,7 +68,7 @@ static int __configure_local(struct sockaddr_storage* storage, const char* addre
 // Windows 10 Insider build 17063 ++ 
 #include <afunix.h>
 
-static int __local_size(void) {
+static int __local_size(const char* address) {
     return sizeof(struct sockaddr_un);
 }
 
@@ -93,7 +110,7 @@ static int init_link_config(struct gracht_link_socket* link, enum gracht_link_ty
             return status;
         }
         domain = AF_LOCAL;
-        size = __local_size();
+        size = __local_size(config->address);
 
         printf("listening at %s\n", config->address);
     } else if (!strcmp(config->type, "inet4")) {
@@ -179,24 +196,6 @@ int waiterd_initialize_server(struct gracht_server_configuration* config, gracht
 #ifdef _WIN32
     // initialize the WSA library
     gracht_link_socket_setup();
-#endif
-
-#if defined(__linux__)
-    status = platform_mkdir("/run/chef/waiterd");
-    if (status) {
-        fprintf(stderr, "waiterd_initialize_server: failed to create /run/chef/waiterd\n");
-        return status;
-    }
-    status = platform_mkdir("/run/chef/waiterd/clients");
-    if (status) {
-        fprintf(stderr, "waiterd_initialize_server: failed to create /run/chef/waiterd/clients\n");
-        return status;
-    }
-    status = platform_chmod("/run/chef/waiterd/clients", 0777);
-    if (status) {
-        fprintf(stderr, "waiterd_initialize_server: failed to set mode for /run/chef/waiterd/clients\n");
-        return status;
-    }
 #endif
 
     status = gracht_server_create(config, serverOut);
